Index robustness + batch overview stats, ETag, lock expiry
- book_meta.json is written atomically (temp file + os.replace): concurrent
  page saves of the same book can no longer truncate or interleave the file
- page sync fetches all PageIndex rows in one query instead of one per page
- index mtimes use integer st_mtime_ns (BigIntegerField, migration 0011) so
  two writes within the same second cannot leave the index stale; the
  book_documents.json fragment comparison keeps its float-second file format
- page delete/rename explicitly invalidate the in-process pcgts cache
- new GET /api/books/overview_stats returns the stats of all readable books
  in one request; both stats endpoints send a content ETag and answer
  If-None-Match with 304
- page edit locks expire after PAGE_EDIT_LOCK_TTL_HOURS (lazily on access
  and in bulk via manage.py release_stale_locks)
- reindex_books --warm-counts precomputes per-page counts
- BooksView pagination parses pageIndex/pageSize as ints
- requirements.txt: add channels-redis and daphne for the ASGI/Redis
  websocket deployment

+++ database/models/book_index.py
@@ -11,7 +11,7 @@ class BookIndex(models.Model):
     """
     name = models.CharField(max_length=255, primary_key=True)
     meta = models.JSONField(default=dict)
-    meta_mtime = models.FloatField(default=0.0)
+    meta_mtime = models.BigIntegerField(default=0)  # st_mtime_ns
 
     # denormalized from `meta` for listing/sorting without JSON access
     display_name = models.CharField(max_length=255, blank=True, default='')
@@ -29,8 +29,8 @@ class PageIndex(models.Model):
     book = models.ForeignKey(BookIndex, on_delete=models.CASCADE, related_name='pages')
     name = models.CharField(max_length=255)
 
-    pcgts_mtime = models.FloatField(default=0.0)      # 0.0 = file absent
-    progress_mtime = models.FloatField(default=0.0)
+    pcgts_mtime = models.BigIntegerField(default=0)   # st_mtime_ns; 0 = file absent
+    progress_mtime = models.BigIntegerField(default=0)
 
     has_symbols = models.BooleanField(default=False)
     verified = models.BooleanField(default=False)
@@ -39,7 +39,7 @@ class PageIndex(models.Model):
 
     # lazily computed symbol/line counts for BookStatsView; null = not yet computed
     counts = models.JSONField(null=True, blank=True)
-    counts_mtime = models.FloatField(default=0.0)
+    counts_mtime = models.BigIntegerField(default=0)
 
     class Meta:
         unique_together = [('book', 'name')]
@@ -50,7 +50,7 @@ class BookDocumentsIndex(models.Model):
     """Parse-avoiding mirror of book_documents.json (which stays the file of record)."""
     book = models.OneToOneField(BookIndex, on_delete=models.CASCADE, related_name='documents_index')
     documents = models.JSONField(null=True)
-    file_mtime = models.FloatField(default=0.0)
+    file_mtime = models.BigIntegerField(default=0)    # st_mtime_ns
 
 
 class PageEditLock(models.Model):
